Fixing (justified) compilation warning on W3I6MV.  1uL is not a 64-bit value.
Copied from Perforce
 Change: 190722
 ServerID: perforce.ravenbrook.com

+++ code/testlib.c
@@ -238,8 +238,8 @@ size_t rnd_align(size_t min, size_t max)
   unsigned log2min = sizelog2(min);
   unsigned log2max = sizelog2(max);
   Insist(min <= max);
-  Insist(1uL << log2min == min);
-  Insist(1uL << log2max == max);
+  Insist((size_t)1 << log2min == min);
+  Insist((size_t)1 << log2max == max);
   if (log2min < log2max)
     return min << (rnd() % (log2max - log2min + 1));
   else
